Delete Format tests which do nothing for coverage
These tests were in the wrong place (and repeated).

Tests under FormatTest should be concerned with ONLY (1) the positions
of the header and body, (2) the body's length encoded into the header,
(3) too small a binary to even have a header, (4) a binary longer than
that for just the one STUN message, (5) related error/exceptions.