Narrow `Client.ip` type to IPv4 IPs
Base TURN RFC requires that we work on IPv4 addresses only.
Unfortunately there might be the case when someone would like
to send Binding request only, which is not restricted to IPv4.
However we can't know that before starting the client, so it's
safer to narrow down the spec to IPv4.